+++ 4.SharedMap/include/shalloc.h
