+++ 4.SharedMap/include/shmemerr.h
